+++ src/main/java/com/ksssss/springframework/beans/factory/support/BeanDefinitionRegistry.java
@@ -16,4 +16,6 @@ public interface BeanDefinitionRegistry {
     void registerBeanDefinition(String beanName, BeanDefinition beanDefinition);
 
     int getBeanDefinitionCount();
+
+    boolean containsBeanDefinition(String beanName);
 }